test: make static-cache e2e test self-contained; slim shared module to probe bridge
Per the tightened consolidation DoD, the e2e Flash-dispatch test no longer
depends on tests/_static_cache_support.py: its e2e-only helpers (tiny-model
builder, cache feeds, and the VERBOSE Flash-dispatch capture) are inlined, and
it imports the canonical probe from mobius._testing.ort_capabilities
(CUDA_AVAILABLE, supports_static_cache_flash).

tests/_static_cache_support.py is reduced to a minimal back-compat bridge that
only re-exports the canonical probe under its historical name, so the
not-yet-migrated parity test keeps collecting.  The bridge is deleted in a
follow-up once the parity test imports the canonical module directly.

Verified on A100 (ORT 1.28.0 w/ #28958): e2e 2/2 and parity 3/3 pass; isolated
ruff check + format clean.

Agent-signed-off: Developer (ed516084) [claude-opus-4.8 via copilot]
